refactor. file structure
